requirements.txt
openpyxl
streamlit
pandas
numpy
scikit-learn
matplotlib
shap
mysql-connector-python

